refactor: 웹 페이지로 연결해주는 메서드들을 클래스로 분리
